requirements.txt
# Core Framework
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1

# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# SAM (Segment Anything Model)
segment-anything==1.0
git+https://github.com/facebookresearch/segment-anything.git

# Segmentation Models
segmentation-models-pytorch==0.3.3
ultralytics==8.0.195

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-image==0.21.0
scipy==1.11.2

# Data Augmentation
albumentations==1.3.1
imgaug==0.4.0

# Dataset Formats
pycocotools==2.0.6
labelme==5.2.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Reporting
openpyxl==3.1.2
reportlab==4.0.4
fpdf==1.7.2

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
colorlog==6.7.0
requests==2.31.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-qt==4.2.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

# Jupyter (optional for development)
jupyter==1.0.0
ipykernel==6.25.1
